refactor: update testing strategy for performance optimization and fix import errors
## Testing Strategy Updates
- Reduce coverage target from 90% to 60% for practical quality assurance
- Set test execution target from 2+ minutes to under 1 minute
- Target test count reduction from 240 to ~150 tests (remove redundant tests)
- Prioritize development efficiency while maintaining quality

## Import Error Fixes
- Fix all layout_old import paths in pages (Dashboard, Income, Expense, History, Settings)
- Rename history/common_old to history/common and update imports
- Fix balance feature export path from components_old to components
- Update legacy AmountInput imports to use new @/components/ui path

## Performance Optimizations
- Add React.memo and useMemo to TransactionForm for render optimization
- Add React.memo and useCallback to AppNavigation for state optimization
- Optimize useAmountInput hook with memoization patterns

## Documentation Updates
- Update testing strategy priorities: test performance > coverage numbers
- Adjust quality gates for realistic development workflow
- E2E testing planned for backend integration with expense input flow priority
- React 19 act() warnings deprioritized (functional quality maintained)
