+++ docs-src/howto/testing-efficient.md
@@ -6,7 +6,7 @@
 
 ### 全テスト実行
 ```bash
-make test-frontend        # 240テスト、25スイート実行
+make test-frontend        # 目標: 150テスト前後、1分以内（現在軽量化作業中）
 ```
 
 ### 特定ファイルのみ実行（推奨）
@@ -26,10 +26,10 @@ make test-file FILE="Controlled"  # フォーム関連
 
 ## ⚡ パフォーマンス最適化
 
-### Jest設定最適化済み
-- **timeout延長**: 15秒（重いテスト対応）
-- **act()警告解決**: React状態更新の適切なラップ
-- **Ripple無効化**: MUIアニメーション無効化で高速化
+### Jest設定最適化済み・軽量化方針
+- **timeout設定**: 30秒（重いテスト対応、但し軽量化で不要になる予定）
+- **React 19 act()警告**: 簡単な修正のみ実施、大規模修正は後回し
+- **テスト数削減**: 重複テスト・過剰詳細テストを削除して高速化
 
 ### 効率的な開発フロー
 1. **コード修正**
@@ -59,15 +59,21 @@ make test-coverage-open   # ブラウザでカバレッジ表示
 
 ### よくある問題
 
-#### テストが遅い
+#### テストが遅い（最優先対応課題）
 ```bash
 # 特定ファイルのみ実行
 make test-file FILE=問題のファイル.test.tsx
+
+# 2024年12月優先: テスト軽量化で根本解決
+# - 重複テストの特定・削除
+# - 過剰詳細テストの削減
+# - 実行昨1分以内を目標とする
 ```
 
-#### act() 警告が出る
-- 既存コードで解決済み
-- 新規コンポーネントは既存パターンを参照
+#### React 19 act() 警告が出る（低優先対応）
+- 現状: 大量のコンソール出力があるが機能的には問題なし
+- 方針: 簡単に修正できる部分のみ対応、大規模修正は後回し
+- 優先度: テスト実行時間短縮 > act()警告完全解決
 
 #### TimeoutError
 - Jest timeout 15秒に設定済み
