+++ docs-src/testing/README.md
@@ -22,15 +22,17 @@ make test-frontend                    # 全単体テスト実行
 make test-file FILE=AmountInput.test.tsx  # 特定ファイル実行
 ```
 
-**テスト範囲**:
-- **UIコンポーネント**: プロパティ・レンダリング・インタラクション
-- **カスタムフック**: 状態変化・副作用・戻り値
-- **ライブラリ関数**: 入力→出力・エッジケース・エラーハンドリング
+**テスト範囲（軽量化方針）**:
+- **UIコンポーネント**: 主要プロパティ・基本動作のみ（内部実装詳細は除外）
+- **カスタムフック**: 主要状態変化・重要な副作用のみ
+- **ライブラリ関数**: 基本的な入力→出力・重大エラーケースのみ
+- **削除対象**: 重複テスト・過剰詳細テスト・エッジケース過多
 
 **品質基準**:
-- **カバレッジ**: Statements 90%以上、Branches 85%以上
-- **実行時間**: 1ファイル10秒以内、全体2分以内
+- **カバレッジ**: Statements 60%以上、Branches 50%以上（実用的品質重視）
+- **実行時間**: 1ファイル5秒以内、全体1分以内（開発効率優先）
 - **安定性**: Flaky test 0%（再実行で結果変動なし）
+- **テスト数**: 150テスト前後（過剰テスト削減、必要十分な品質保証）
 
 ### 2. 統合テスト (Integration Tests)
 **対象**: コンポーネント間連携・ルーティング・データフロー
@@ -58,10 +60,10 @@ make test-frontend
 make test-e2e                         # Playwright/Cypress
 ```
 
-**テスト範囲（計画）**:
-- **主要フロー**: 支出登録→履歴確認→削除
-- **認証フロー**: ログイン→操作→ログアウト
-- **レスポンシブ**: デスクトップ・タブレット・モバイル
+**テスト範囲（計画・バックエンド連携時）**:
+- **最優先フロー**: 支出入力（金額入力→バリデーション→保存→履歴表示）
+- **次優先フロー**: 収入入力・残高計算
+- **将来対応**: 認証フロー・レスポンシブ対応
 
 ### 4. 契約テスト (Contract Tests)
 **対象**: Frontend↔Backend API仕様の整合性
@@ -78,19 +80,20 @@ make contract-test                    # Pact/OpenAPI Testing
 
 ## 📊 現在の品質状況
 
-### テスト実行状況
+### テスト実行状況・軽量化目標
 ```
-✅ 単体テスト: 240テスト、25スイート （100%通過）
-✅ 統合テスト: ルーティング・データフロー （包含済み）
-🔄 E2Eテスト: 未実装（将来対応）
-🔄 契約テスト: 未実装（OpenAPI準備中）
+⚠️ 現状: 240テスト、25スイート （実行昨2分超過で改善必要）
+🎯 目標: 150テスト前後、18スイート （実行昨1分以内）
+✅ 統合テスト: ルーティング・支出入力フロー （重要部分のみ）
+🔄 E2Eテスト: バックエンド連携時に支出入力優先導入
+❌ 契約テスト: 現在は非対応
 ```
 
-### カバレッジメトリクス
-- **Statements**: 高カバレッジ維持中
-- **Branches**: 条件分岐網羅
-- **Functions**: 全関数テスト済み
-- **Lines**: ソースコード行カバレッジ
+### カバレッジ目標（現実的調整）
+- **Statements**: 60%以上（旧目標90%から実用的に調整）
+- **Branches**: 50%以上（主要条件分岐のみ）
+- **Functions**: 重要関数のみテスト（内部ヘルパー関数は除外）
+- **品質方針**: 動作保証 > カバレッジ数値（実際の品質を優先）
 
 ## ⚡ 効率的なテスト実行戦略
 
@@ -150,11 +153,12 @@ make quality-check-frontend
 
 ## 📋 品質ゲート基準
 
-### リリース前必須チェック
-- [ ] **全テスト通過**: 240テスト、0エラー、0警告
-- [ ] **カバレッジ基準**: Statements 90%以上
-- [ ] **パフォーマンス**: 全テスト2分以内完了
+### リリース前必須チェック（軽量化対応）
+- [ ] **全テスト通過**: 150テスト前後、0エラー、0重大警告
+- [ ] **カバレッジ基準**: Statements 60%以上（動作保証があること）
+- [ ] **パフォーマンス**: 全テスト1分以内完了（開発効率優先）
 - [ ] **安定性**: 3回連続実行で同一結果
+- [ ] **主要機能**: 支出入力フロー動作確認
 
 ### コード品質統合チェック
 ```bash
@@ -170,10 +174,13 @@ make quality-check-frontend
 
 ### よくある問題と解決策
 
-#### 1. テスト実行が遅い
+#### 1. テスト実行が遅い（最優先対応）
 ```bash
 # 特定ファイルのみ実行
 make test-file FILE=遅いテスト.test.tsx
+
+# 重複テストの特定・削除を2024年12月優先実施
+# 過剰詳細テストの削除で実行昨1分以内を目指す
 
 # キャッシュクリア
 cd frontend && npm test -- --clearCache
@@ -207,17 +214,19 @@ jest.mock('@/hooks/useMoney', () => ({
 
 ## 🚀 テスト戦略の進化
 
-### 短期計画（次四半期）
-1. **E2Eテスト導入**: Playwright環境構築
-2. **ビジュアル回帰テスト**: Storybook + Chromatic
-3. **契約テスト準備**: OpenAPI仕様完成→テスト自動化
+### 短期計画（次四半期・優先度順）
+1. **テスト軽量化**: 重複テスト削除で実行昨1分以内を実現（最優先）
+2. **E2Eテスト準備**: バックエンド連携時に支出入力フローのPlaywright導入
+3. **品質基準調整**: カバレッジ60%目標への実運用調整
+4. **非対応**: ビジュアル回帰テスト・契約テスト
 
-### 中期計画（半年〜1年）
-1. **パフォーマンステスト**: Lighthouse CI統合
-2. **アクセシビリティテスト**: axe-core自動化
-3. **セキュリティテスト**: 脆弱性スキャン統合
+### 中期計画（半年〜1年・選択的導入）
+1. **パフォーマンステスト**: 重要コンポーネントのみ部分的導入検討
+2. **非対応**: アクセシビリティテスト
+3. **非対応**: セキュリティテスト
 
-### 品質文化醸成
-- **テストファースト**: TDD推進・レッドグリーンリファクター
-- **品質メトリクス**: ダッシュボード化・継続監視
-- **学習・改善**: 定期的なテスト戦略レビュー・技術向上
+### 品質文化醸成（実用的アプローチ）
+- **効率重視**: 開発速度と品質のバランスを重視したテスト戦略
+- **動作保証優先**: カバレッジ数値より実際の動作品質を優先
+- **適切な範囲**: 必要十分なテストのみで品質を保証、過剰テストを回避
+- **定期見直し**: 四半期毎にテスト効率と品質のバランスを再評価
