+++ frontend/src/features/balance/index.ts
@@ -2,4 +2,4 @@
  * 残高表示機能モジュール
  * 家計簿の現在の残高を表示するためのコンポーネントを提供します
  */
-export { default as BalanceDisplay } from './components_old/BalanceDisplay'
+export { default as BalanceDisplay } from './components/BalanceDisplay'
