+++ frontend/src/pages/HistoryPage.tsx
@@ -1,5 +1,5 @@
 import { Box, Chip, Grid, Paper, Typography } from '@mui/material'
-import AppLayout from '@/components/layout_old/AppLayout'
+import AppLayout from '@/components/layout/AppLayout'
 import { ExpenseHistory, IncomeHistory } from '@/features/history'
 import { useBudgetManager } from '@/hooks/useBudgetManager'
 import { formatMoney } from '@/lib/format'
